improved cmake structure
