+++ apps/CMakeLists.txt
@@ -2,9 +2,7 @@
 cmake_minimum_required(VERSION 3.5...3.27)
 
 # Add Build Targets
-
-
-IF (ENABLE_QT AND Qt5Widgets_FOUND)
-add_subdirectory(line_analyzer)
-ENDIF (ENABLE_QT AND Qt5Widgets_FOUND)
+if(ENABLE_QT AND Qt5Widgets_FOUND)
+    add_subdirectory(line_analyzer)
+endif()
