+++ examples/thesis/edge_seg.cpp
@@ -2,6 +2,7 @@
 #include <edge/nfa.hpp>
 #include <edge/nms.hpp>
 #include <edge/threshold.hpp>
+#include <edge/draw.hpp>
 #include <geometry/draw.hpp>
 #include <imgproc/derivative_gradient.hpp>
 #include <eval/eval_app.hpp>
