+++ libs/CMakeLists.txt
@@ -9,7 +9,3 @@ add_subdirectory(edge)
 add_subdirectory(imgproc)
 add_subdirectory(geometry)
 add_subdirectory(phase_cong)
-
-IF (ENABLE_QT AND Qt5Widgets_FOUND)
-    # moved to third-party/qplot and third-party/qplot3d
-ENDIF (ENABLE_QT AND Qt5Widgets_FOUND)
