+++ libs/edge/CMakeLists.txt
@@ -1,21 +1,16 @@
 cmake_minimum_required(VERSION 3.5...3.27)
 
-file(GLOB edge_hpp "include/edge/*.hpp")
-add_library(lib_edge ${LIBRARY_TYPE} ${edge_hpp})
-
-target_include_directories(lib_edge
-    PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include
+# Create edge library using automatic source detection
+# Edge depends on both imgproc and geometry
+le_add_library(lib_edge
+    AUTO_TESTS
+    PUBLIC_DEPS 
+        lib_imgproc
+        lib_geometry
+        ${OpenCV_LIBS}
+    PUBLIC_INCLUDES
+        ${CMAKE_CURRENT_SOURCE_DIR}/include
+        ${OpenCV_INCLUDE_DIR}
+    TEST_DEPS
+        ${CMAKE_THREAD_LIBS_INIT}
 )
-
-target_link_libraries(lib_edge
-    PUBLIC lib_imgproc
-    PUBLIC lib_geometry
-    PUBLIC ${OpenCV_LIBS}
-)
-
-SET_TARGET_PROPERTIES(lib_edge PROPERTIES LINKER_LANGUAGE CXX)
-
-# Add tests if testing is enabled
-if(ENABLE_UNIT_TEST OR BUILD_TESTING)
-    add_subdirectory(tests)
-endif()
