+++ libs/edge/include/edge/draw.hpp
@@ -1,0 +1,76 @@
+/*M///////////////////////////////////////////////////////////////////////////////////////
+// IMPORTANT: READ BEFORE DOWNLOADING, COPYING, INSTALLING OR USING.
+//
+//  By downloading, copying, installing or using the software you agree to this license.
+//  If you do not agree to this license, do not download, install,
+//  copy or use the software.
+//
+//
+//                           License Agreement
+//                For Open Source Computer Vision Library
+//
+// Copyright (C) 2000-2008, Intel Corporation, all rights reserved.
+// Copyright (C) 2008-2011, Willow Garage Inc., all rights reserved.
+// Third party copyrights are property of their respective owners.
+//
+// Redistribution and use in source and binary forms, with or without modification,
+// are permitted provided that the following conditions are met:
+//
+//   * Redistributions of source code must retain the above copyright notice,
+//     this list of conditions and the following disclaimer.
+//
+//   * Redistributions in binary form must reproduce the above copyright notice,
+//     this list of conditions and the following disclaimer in the documentation
+//     and/or other materials provided with the distribution.
+//
+//   * The name of the copyright holders may not be used to endorse or promote products
+//     derived from this software without specific prior written permission.
+//
+// This software is provided by the copyright holders and contributors "as is" and
+// any express or implied warranties, including, but not limited to, the implied
+// warranties of merchantability and fitness for a particular purpose are disclaimed.
+// In no event shall the Intel Corporation or contributors be liable for any direct,
+// indirect, incidental, special, exemplary, or consequential damages
+// (including, but not limited to, procurement of substitute goods or services;
+// loss of use, data, or profits; or business interruption) however caused
+// and on any theory of liability, whether in contract, strict liability,
+// or tort (including negligence or otherwise) arising in any way out of
+// the use of this software, even if advised of the possibility of such damage.
+//
+// C by Benjamin Wassermann
+//M*/
+
+#ifndef _EDGE_DRAW_HPP_
+#define _EDGE_DRAW_HPP_
+#ifdef __cplusplus
+
+#include <opencv2/opencv.hpp>
+#include <opencv2/imgproc/imgproc.hpp>
+#include <opencv2/imgproc/types_c.h>
+#include <edge/edge_segment.hpp>
+#include <utility/range.hpp>
+
+namespace lsfm {
+
+    // Draw functions for edge-specific types
+    inline void drawSegment(cv::Mat& img, const EdgeSegment& seg, const IndexVector& points, const cv::Vec3b& color = cv::Vec3b(0, 0, 255)) {
+        if (img.channels() == 1)
+            cv::cvtColor(img,img,CV_GRAY2RGB);
+        cv::Vec3b *pimg = img.ptr<cv::Vec3b>();
+        for (size_t i = seg.begin(); i != seg.end(); ++i) {
+            pimg[points[i]] = color;
+        }
+    }
+
+    inline void drawSegment(cv::Mat& img, const EdgeSegmentVector& segs, const IndexVector& points, const cv::Vec3b& color = cv::Vec3b(0, 0, 255)) {
+        if (img.channels() == 1)
+            cv::cvtColor(img,img,CV_GRAY2RGB);
+        for_each(segs.begin(),segs.end(),[&](const EdgeSegment& seg) {
+            drawSegment(img,seg,points,color);
+        });
+    }
+
+}
+
+#endif
+#endif
