+++ libs/imgproc/CMakeLists.txt
@@ -1,20 +1,15 @@
 cmake_minimum_required(VERSION 3.5...3.27)
 
-file(GLOB imgproc_hpp "include/imgproc/*.hpp")
-add_library(lib_imgproc ${LIBRARY_TYPE} ${imgproc_hpp})
-
-target_include_directories(lib_imgproc
-    PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include
+# Create imgproc library using automatic source detection
+le_add_library(lib_imgproc
+    AUTO_TESTS
+    PUBLIC_DEPS 
+        lib_utility
+        lib_phase_cong
+        ${OpenCV_LIBS}
+    PUBLIC_INCLUDES
+        ${CMAKE_CURRENT_SOURCE_DIR}/include
+        ${OpenCV_INCLUDE_DIR}
+    TEST_DEPS
+        ${CMAKE_THREAD_LIBS_INIT}
 )
-
-target_link_libraries(lib_imgproc
-    PUBLIC lib_utility
-    PUBLIC ${OpenCV_LIBS}
-)
-
-SET_TARGET_PROPERTIES(lib_imgproc PROPERTIES LINKER_LANGUAGE CXX)
-
-# Add unit tests for this library when enabled
-if(ENABLE_UNIT_TEST)
-  add_subdirectory(tests)
-endif()
